debugg sur noms donnés aux types, application run correctement à ce stade

+++ src/main/java/com/jeromerichard/pdfstream/Service/Interfaces/AlertServiceInt.java
@@ -7,6 +7,7 @@ import com.jeromerichard.pdfstream.Entity.Pdf;
 import com.jeromerichard.pdfstream.Entity.User;
 import com.jeromerichard.pdfstream.Exception.EmptyListException;
 import com.jeromerichard.pdfstream.Exception.NotFoundException;
+import org.hibernate.validator.constraints.ru.INN;
 
 import java.util.Set;
 
@@ -14,9 +15,9 @@ public interface AlertServiceInt {
     // CRUD JPA
     public Alert saveAlert(AlertDTODown alert);
     public Set<Alert> getAllAlerts() throws EmptyListException;
-    public Alert getAlertById(Long id) throws NotFoundException;
-    public Alert updateAlert(Long id, AlertDTODown alert) throws NotFoundException;
-    public void deleteAlert(Long id) throws NotFoundException;
+    public Alert getAlertById(Integer id) throws NotFoundException;
+    public Alert updateAlert(Integer id, AlertDTODown alert) throws NotFoundException;
+    public void deleteAlert(Integer id) throws NotFoundException;
     // FOREIGN KEY
     public Set<Alert> findByUser(User userId) throws EmptyListException;
     public Set<Alert> findByArticle(Article articleId) throws EmptyListException;
